lib/game: move figures into lib/game for reuse elsewhere
